pyproject.toml
[tool.poetry]
name = "asegurahoy"
version = "0.1.0"
description = ""
authors = ["MiqueasAguirreANX <70060960+MiqueasAguirreANX@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.3"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
